+++ Aelia's Apprentice/Assets/Scripts/Player.cs
@@ -2,26 +2,26 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    [SerializeField] private GameObject projectilePrefab;
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    [SerializeField] private GameObject spellPrefab;
+
     void Start()
     {
-        if (projectilePrefab == null)
-            Debug.LogError(gameObject + " needs to have a projectile prefab");
+        if (spellPrefab == null)
+            Debug.LogError(gameObject + " needs to have a spell prefab");
     }
 
-    // Update is called once per frame
     void Update()
     {
         if (Input.GetKeyDown(KeyCode.Space))
         {
             Vector2 spawnPos = transform.position;
-            Vector2 direction = Vector2.up;
-            float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
-            Quaternion rotation = Quaternion.Euler(0, 0, angle);
+            //Vector2 direction = Vector2.up;
+            //float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
+            //Quaternion rotation = Quaternion.Euler(0, 0, angle);
+            Quaternion rotation = Quaternion.identity;
 
-            GameObject projectileObj = Instantiate(projectilePrefab, spawnPos, rotation);
-            projectileObj.GetComponent<Projectile>().Direction = direction;
+            GameObject projectileObj = Instantiate(spellPrefab, spawnPos, rotation);
+            //projectileObj.GetComponent<Projectile>().Direction = direction;
         }
     }
 }