+++ Aelia's Apprentice/Assets/Scripts/Projectile.cs
@@ -1,12 +1,9 @@
 using UnityEngine;
 
-public class Projectile : MonoBehaviour
+public class Projectile : HitResponder
 {
     // Reference to the Rigidbody2D for movement
     private Rigidbody2D Rb { get; set; }
-
-    // Reference to the Hitbox component used for detecting hits
-    private Hitbox Hitbox { get; set; }
 
     // Direction the projectile moves in
     public Vector2 Direction { get; set; }
@@ -14,11 +11,11 @@ public class Projectile : MonoBehaviour
     // Speed of the projectile in units per second
     private float Speed { get; set; }
 
-    // Tracks whether the projectile has already hit something
-    private bool HasHit { get; set; } = false;
-
-    void Start()
+    protected override void Start()
     {
+        // Call HitResponder.Start() to ensure that setup runs first
+        base.Start();
+
         Rb = GetComponent<Rigidbody2D>();
 
         if (Rb == null)
@@ -26,14 +23,6 @@ public class Projectile : MonoBehaviour
 
         Direction = Vector2.up;
         Speed = 5f;
-
-        Hitbox = GetComponentInChildren<Hitbox>();
-
-        // If this does have a hitbox, subscribe to its hitboxHit
-        if (Hitbox != null)
-            Hitbox.hitboxHit += OnHit;
-        else
-            Debug.LogError(gameObject + " needs a Hitbox component");
     }
 
     // Update is called once per frame
@@ -41,17 +30,5 @@ public class Projectile : MonoBehaviour
     {
         // Move the projectile forward based on direction and speed
         Rb.MovePosition(Rb.position + Direction * Speed * Time.fixedDeltaTime);
-    }
-
-    private void OnHit(Collider2D hitboxCollider)
-    {
-        // Exit early if this has already hit something, otherwise set that it has hit
-        if (HasHit) return;
-        HasHit = true;
-
-        Debug.Log(gameObject + ": My hitbox has hit something");
-
-        Enemy enemy = hitboxCollider.gameObject.GetComponentInParent<Enemy>();
-        enemy.TakeDamage();
     }
 }